+++ cli/dev-server/web-server/files-map.ts
@@ -1,17 +1,26 @@
+import path from 'path';
+
 import { CompilerResult } from './../../../builder/builders/compilers/definitions';
 import { IDynamicCompilerResult } from './../compilation-map';
 
 export class FilesMap extends Map<string, CompilerResult> {
   
   // TODO: Could definetly be smarter
-  resolve(element: string): CompilerResult | undefined {
-    if (this.get(element)) { return this.get(element); }
-    if (this.get(element + '.html')) { return this.get(element + '.html'); }
-    if (this.get(element + 'index.html')) { return this.get(element + 'index.html'); }
+  resolve(element: string = ''): CompilerResult | undefined {
+    const n = this.resolveName(element);
+    if (n) { return this.get(n); }
   }
 
-  has(element: string) {
-    return !!this.resolve(element);
+  resolveName(element: string = ''): string | undefined {
+    if (super.has(element)) { return element; }
+    if (!path.extname(element)) {
+      if (super.has(element + '.html')) { return element + '.html'; }
+      if (super.has(element + 'index.html')) { return element + 'index.html'; }
+    }
+  }
+
+  has(element: string = '') {
+    return !!this.resolveName(element);
   }
 
   static fromDynamic(data: Map<string, IDynamicCompilerResult>): FilesMap {